Safari-609.1.11.1/LayoutTests/svg/dom/SVGDocument-createEvent.html

<!DOCTYPE html>
<html>
<body>
<script src="../../resources/js-test-pre.js"></script>
<script>
description("Test SVGDocument.createEvent() operation.");

var svgDocument = document.implementation.createDocument("http://www.w3.org/2000/svg", "svg");
shouldThrow("svgDocument.createEvent()", "'TypeError: Not enough arguments'");
shouldNotThrow("event = svgDocument.createEvent('mouseevent')");
shouldBe("event.__proto__", "MouseEvent.prototype");

</script>
<script src="../../resources/js-test-post.js"></script>
</body>
</html>
